+++ frontend/src/pages/Methodology/methodologyContent/PdohDefinitions.tsx
@@ -1,6 +1,5 @@
 import { methodologyTableDefinitions } from './MethodologyTopicDefinitions'
 import { conditionVariableDefinitions } from './ConditionVariableDefinitions'
-import { dataSourceMetadataMap } from '../../../data/config/MetadataMap'
 
 export const pdohDefinitionsArray = [
   {
@@ -73,14 +72,4 @@ export const pdohDefinitionsArray = [
       },
     ],
   },
-]
-
-export const pdohDataSources = [
-  dataSourceMetadataMap.bjs,
-  dataSourceMetadataMap.acs,
-  dataSourceMetadataMap.decia_2020_territory_population,
-  dataSourceMetadataMap.ahr,
-  dataSourceMetadataMap.vera,
-  dataSourceMetadataMap.cawp,
-  dataSourceMetadataMap.the_unitedstates_project,
 ]
